Implemented export changes to a file
