+++ EwidencjaWSK/Views/Admin/Audit.cshtml
@@ -23,6 +23,13 @@
                 <input asp-for="To" class="form-control" />
                 <span asp-validation-for="To" class="text-danger"></span>
             </div>
+
+
+            <div class="form-group form-check">
+                <label class="form-check-label">
+                    <input class="form-check-input" asp-for="ExportToFile" /> @Html.DisplayNameFor(model => model.ExportToFile)
+                </label>
+            </div>
             <div class="form-group">
                 <input asp-action="Changes" type="submit" value="Raport zmian" class="btn btn-primary" />
             </div>
